Don't store newly allocated nodes in `hsRef`.
When `hsRef` is constructed from a pointer, it increments the reference
count. Originally, the `NetVaultNode` reference count started at zero,
so this was fine. However, when the various reference counting bases
were consolidated, `NetVaultNode`'s count became one. This means that
any instance of `hsRef<NetVaultNode> ref = new NetVaultNode;` would have
a reference count of two, causing the object to be leaked.